Moves TestPrintUnstructuredObject to tableprinter_test.go
